feat: add data splitting step (#53)
* feat(pipeline): :sparkles: Added named inputs and outputs to PipelineStep.

The PipelineStep now requires inputs and outputs names which will modify a single DataContainer throughout the entire Pipeline. The DataContainer data attribute is now a dictionary and allows steps to specifically select which data should be sent to the wrapped class executing the logic. This allows for splitting data to multiple DataFrames and specifically selecting one of them in future steps for instance.

Classes inheriting from PipelineSteps will need to be appended with _num_inputs and _num_outputs. The initialization of the IngestStep and ConverterStep needs to include the new parameters inputs and outputs.

* feat(cache): :sparkles: Added fingerprinter for Vaex DataFrames.

* feat(cache): :sparkles: Cache not supports both single and multiple outputs from cached methods.

If a method returns multiple objects in a list or tuple they will be individually cached as files on the format of hash_index.suffix. When multiple files of a single hash is found they will be loaded in the correct order.

* feat(data splitting): :sparkles: Added data splitting classes and pipeline step which splits a dataframe in two.

The RandomDataSplitter support random (shuffled, stratified) splitting according to ratios. The ExpressionDataSplitter splits data based on boolean string expressions.

The caching mechanism is updated to accomodate for multiple file outputs.

* fix: :bug: Updated ignore list for library stubs to also include scikit-learn.

* fix: :bug: Forced datetime columns option removed from convert step.

---------

Co-authored-by: Erik Båvenstrand <erik.bavenstrand@klarna.com>